Moving Galley source to dynamic package.
This is in preparation for #10995. Trying to do this move in order
to preserve history.
