+++ galley/pkg/source/kube/dynamic/listener.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kube
+package dynamic
 
 import (
 	"fmt"
